Fix CCE
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@589312 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/product/src/org/ofbiz/shipment/shipment/ShipmentServices.java
@@ -380,7 +380,7 @@ public class ShipmentServices {
                 }
 
                 // add the feature quantities
-                Double quantity = (Double) itemMap.get("quantity");
+                BigDecimal quantity = (BigDecimal) itemMap.get("quantity");
                 Set featureSet = (Set) itemMap.get("featureSet");
                 if (featureSet != null && featureSet.size() > 0) {
                     Iterator fi = featureSet.iterator();
